--- data/Messages.txt
@@ -1,14 +1,0 @@
-[
-   {
-      "correo": "correo@correo.com",
-      "mensaje": "mensaje 1",
-      "createdAt": "25/10/2022 18:51:0",
-      "id": 1
-   },
-   {
-      "correo": "correo@correo.com",
-      "mensaje": "mensaje 2",
-      "createdAt": "25/10/2022 18:52:37",
-      "id": 2
-   }
-]
