--- models/message/message.model.js
@@ -1,6 +1,0 @@
-
-const Contenedor = require('../Contenedor.js')
-
-const Messages = new Contenedor('Messages')
-
-module.exports = Messages
