--- models/product/product.model.js
@@ -1,5 +1,0 @@
-const Contenedor = require('../../models/Contenedor')
-
-const Products = new Contenedor('Products')
-
-module.exports = Products
