--- public/assets/templates/chat.template.hbs
@@ -1,12 +1,0 @@
-<h2 id="chat__title">Chat</h2>
-<div id="chat_dashboard">
-{{#if messages.length}}
-    {{#each messages}}
-    <div class="chat-message">
-        <p class="chat-message-text">{{this.correo}}</p>
-        <p class="chat-message-text">{{this.mensaje}}</p>
-        <p class="chat-message-date">{{this.createdAt}}</p>
-    </div>
-    {{/each}}
-{{/if}}
-</div>
