--- public/assets/templates/product.template.hbs
@@ -1,26 +1,0 @@
-<h2 id="products__title">Productos</h2>
-<div id="products_dashboard">
-    <table class="product-list">
-        <tr>
-            <th>Título</th>
-            <th>Precio</th>
-            <th>Miniatura</th>
-        </tr>
-{{#if products.length}}
-    {{#each products}}
-    <div class="product">
-        <tr>
-            <div class="product__info">
-            <td class='product__title'>{{this.title}}</td>
-            <td class='product__price'>{{this.price}}</td>
-            <td><img height="72px" width="72px" src={{this.thumbnail}} class="product__img" /></td>
-            </div>
-        </tr>
-    </div>
-    {{/each}}
-     </table>
-  {{else}}
-  <p id="products_empty">No posee productos ingresados</p>
-{{/if}}
-</div>
-
