--- views/form-products.hbs
@@ -1,8 +1,0 @@
-<form action="/productos" method="post">
-        <input type="text" name="title" placeholder="Titulo" required>
-        <input type="number" name="price" placeholder="0.00" required>
-        <input type="url" name="thumbnail" placeholder="thumbnail" required>
-        <input type="submit" value="Agregar">
-</form>
-
-<a href="/productos"><button>Ir a productos</button></a>
