--- views/layouts/main.hbs
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <h1 class="title">Bienvenidos a productos</h1>
-    {{{body}}}
-</body>
-</html>
-
-<style>
-    .title {
-        text-align: center;
-    }
-</style>
